+++ tests/manage_config/mod.rs
@@ -1,0 +1,1 @@
+pub mod manage_config_int_test;
