+++ tests/mod.rs
@@ -1,0 +1,10 @@
+extern crate assert_cli;
+
+// Sub modules
+pub mod common;
+
+pub mod manage_config;
+
+pub mod proxy_config;
+
+pub mod repository_config;
