+++ tests/proxy_config/mod.rs
@@ -1,0 +1,1 @@
+pub mod proxy_config_int_test;
