+++ tests/repository_config/mod.rs
@@ -1,0 +1,1 @@
+pub mod repository_config_int_test;
